refactor!: Move file_sink to `sinks/` subfolder
