+++ libs/deeplog/CMakeLists.txt
@@ -78,12 +78,12 @@ dplx_target_sources(deeplog
         dlog/source/log_record_port
         dlog/source/span_scope
 
+        dlog/sinks/file_sink
         dlog/sinks/sink_frontend
 
         dlog/attributes
         dlog/attribute_transmorpher
         dlog/log_fabric
-        dlog/sink
 
         dlog/bus/buffer_bus
         dlog/bus/mpsc_bus
