+++ libs/deeplog/src/dplx/dlog/record_container.hpp
@@ -22,12 +22,13 @@
 #include <dplx/dp/codecs/core.hpp>
 #include <dplx/dp/codecs/std-string.hpp>
 #include <dplx/dp/items/parse_core.hpp>
+#include <dplx/dp/items/skip_item.hpp>
 
 #include <dplx/dlog/argument_transmorpher_fmt.hpp>
 #include <dplx/dlog/attribute_transmorpher.hpp>
 #include <dplx/dlog/attributes.hpp>
 #include <dplx/dlog/core/strong_types.hpp>
-#include <dplx/dlog/sink.hpp>
+#include <dplx/dlog/sinks/file_sink.hpp>
 
 namespace dplx::dlog
 {
