--- libs/deeplog/src/dplx/dlog/sink.cpp
@@ -1,10 +1,0 @@
-
-// Copyright Henrik Steffen Gaßmann 2021
-//
-// Distributed under the Boost Software License, Version 1.0.
-//         (See accompanying file LICENSE or copy at
-//           https://www.boost.org/LICENSE_1_0.txt)
-
-#include <dplx/dlog/sink.hpp>
-
-DPLX_DLOG_DEFINE_AUTO_OBJECT_CODEC(::dplx::dlog::file_info)
